requirements.txt
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
ta==0.11.0
scikit-learn==1.4.0
tensorflow==2.15.0
xgboost==2.0.3
transformers==4.37.2
torch==2.2.0
vaderSentiment==3.3.2
textblob==0.17.1
backtrader==1.9.78.123
streamlit==1.31.0
python-dotenv==1.0.1
matplotlib==3.8.2
plotly==5.18.0
requests==2.31.0
streamlit-lottie==0.0.3
